Integrate change 125470 from Main to FRANKLIN:

	----------
	Change 125470 by tim@timsmac2 on 2008/12/03 16:09:50

		bug: none

		cleanup

http://bugzilla.zimbra.com/show_bug.cgi?id=20945
http://bugzilla.zimbra.com/show_bug.cgi?id=29483

Copied from Perforce
 Change: 140245

+++ src/java/com/zimbra/cs/mailbox/Mailbox.java
@@ -6862,7 +6862,7 @@ public class Mailbox {
                 List<Integer> deferredTagsToSet = new ArrayList<Integer>();
                 
                 try {
-                    if (mMailboxIndex != null)
+                    if (mMailboxIndex != null && (mCurrentChange.indexItemsToDelete.size() > 0))
                         mMailboxIndex.deleteDocuments(mCurrentChange.indexItemsToDelete);
                 } catch (IOException e) {
                     if (ZimbraLog.index.isDebugEnabled())
@@ -6968,7 +6968,8 @@ public class Mailbox {
                     ZimbraLog.index.warn("Possibly corrupt index: "+numIndexingExceptions+" indexing operations failed in mailbox transaction.");
                     mLastIndexingFailureTimestamp = System.currentTimeMillis();
                 }
-            }
+                
+            } // if indexing needed
             
             // 4. Commit the main transaction in database.
             if (conn != null) {
